backend/requirements.txt
# Django and DRF
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# Cloudinary storage
django-cloudinary-storage==0.3.0
cloudinary==1.39.0

# Audit and history
django-simple-history==3.4.0

# Configuration
django-environ==0.11.2

# API Documentation
drf-spectacular==0.27.1

# CORS headers
django-cors-headers==4.3.1

# OTP
django-otp==1.3.0
qrcode==7.4.2

# Celery for async tasks
celery==5.3.6
redis==5.0.1

# Encryption
cryptography==42.0.2

# Rate limiting
django-ratelimit==4.1.0

# Production server
gunicorn==21.2.0

# Image processing
Pillow==10.2.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1

